// src/telegram/enums/action_session.enum.ts
export enum Action {
  LOGIN = 'Login',
  ARRIVAL = 'Arrival',
  RECORD_REMAINING = 'Record Remaining Stock',
  CLOSE = 'Close',
}

export enum SessionStage {
  LOGIN_PROMPT = 'LOGIN_PROMPT',
  LOGGED_IN = 'LOGGED_IN',
  USERNAME = 'USERNAME',
  PASSWORD = 'PASSWORD',
  CHOOSING_CATEGORY = 'CHOOSING_CATEGORY',
  CHOOSING_PRODUCT = 'CHOOSING_PRODUCT',
  ENTERING_QUANTITY = 'ENTERING_QUANTITY',
  VIEW_ARRIVAL = 'VIEW_ARRIVAL',
  ENTERING_REMAINING_STOCK = 'ENTERING_REMAINING_STOCK',
  RECORD_REMAINING = 'RECORD_REMAINING',
}